+++ Core/Runtime/KegEvent.cs
@@ -1,0 +1,8 @@
+﻿namespace Kegstand
+{
+    public enum KegEvent
+    {
+        Filled,
+        Emptied,
+    }
+}
