+++ Core/Runtime/TimedEvent.cs
@@ -1,0 +1,9 @@
+﻿namespace Kegstand
+{
+    public class TimedEvent
+    {
+        public float Time { get; set; }
+        public int Index { get; set; }
+        public KegEvent Type { get; set; }
+    }
+}
